Create Snake and Ladder Problem in Graph.py
Given a snake and ladder board, find the minimum number of dice throws required to reach the destination or last cell from the source or 1st cell. Basically, the player has total control over the outcome of the dice throw and wants to find out the minimum number of throws required to reach the last cell.
If the player reaches a cell that is the base of a ladder, the player has to climb up that ladder, and if they reach a cell that is the mouth of the snake, and has to go down to the tail of the snake without a dice throw.